+++ vscode/webviews/components/sidebar.svelte
@@ -135,5 +135,5 @@
 {/if}
 
 <div class="row-pt-2">
-<a href="https://github.com/aiops/autologger">Join our [community]</a>
+<a href="https://github.com/aiops/log-writer">Join our [community]</a>
 </div>