+++ vscode/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "autologger" extension will be documented in this file.
+All notable changes to the "log-writer" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
@@ -8,7 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [1.0.0] - 05.11.2022
 ### Added
-- Autologging enabled for VSCode 
+- Log writing enabled for VSCode 
 - Enabling logging recommendations for PYTHON
 
 ### Changed